# api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Don't change title and version, these are used in the code generation and should be consistent with the package.json version. It affects import paths and generated client code. You can update the description to reflect the current state of the API, but keep title and version as is for consistency.
  title: Api
  version: 0.1.0
  description: Rayo AI Personal Finance API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List all transactions
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: description
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [income, expense]
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: amount
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
        "400":
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionBody"
      responses:
        "201":
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTransactions
      tags: [transactions]
      summary: Delete multiple transactions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteTransactionsBody"
      responses:
        "204":
          description: Deleted
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /transactions/{id}:
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /savings:
    get:
      operationId: listSavingsGoals
      tags: [savings]
      summary: List savings goals
      responses:
        "200":
          description: List of savings goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavingsGoal"
    post:
      operationId: createSavingsGoal
      tags: [savings]
      summary: Create a savings goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSavingsGoalBody"
      responses:
        "201":
          description: Savings goal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavingsGoal"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /savings/{id}:
    patch:
      operationId: updateSavingsGoal
      tags: [savings]
      summary: Update a savings goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSavingsGoalBody"
      responses:
        "200":
          description: Updated savings goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavingsGoal"
        "404":
          description: Savings goal not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteSavingsGoal
      tags: [savings]
      summary: Delete a savings goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Savings goal not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
        category:
          type: string
        description:
          type: string
        date:
          type: string
        createdAt:
          type: string
      required:
        - id
        - userId
        - type
        - amount
        - category
        - date
        - createdAt
    CreateTransactionBody:
      type: object
      properties:
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
          minimum: 0.01
        category:
          type: string
          minLength: 1
        description:
          type: string
        date:
          type: string
          minLength: 1
      required:
        - type
        - amount
        - category
        - date
    DeleteTransactionsBody:
      type: object
      properties:
        ids:
          type: array
          items:
            type: integer
          minItems: 1
      required:
        - ids
...
    SavingsGoal:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        targetAmount:
          type: number
        currentAmount:
          type: number
        deadline:
          type: string
        percentComplete:
          type: number
      required:
        - id
        - userId
        - name
        - targetAmount
        - currentAmount
        - deadline
        - percentComplete
    CreateSavingsGoalBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        targetAmount:
          type: number
          minimum: 0.01
        currentAmount:
          type: number
          minimum: 0
        deadline:
          type: string
          minLength: 1
      required:
        - name
        - targetAmount
        - deadline
    UpdateSavingsGoalBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        targetAmount:
          type: number
          minimum: 0.01
        currentAmount:
          type: number
          minimum: 0
        deadline:
          type: string
      required: []
...