# validation/generated/types/createTransactionBody.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Api
 * Rayo AI Personal Finance API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTransactionBodyType } from './createTransactionBodyType';

export interface CreateTransactionBody {
  type: CreateTransactionBodyType;
  /** @minimum 0.01 */
  amount: number;
  /** @minLength 1 */
  category: string;
  description?: string;
  /** @minLength 1 */
  date: string;
}