# validation/generated/types/deleteTransactionsBody.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Api
 * Rayo AI Personal Finance API
 * OpenAPI spec version: 0.1.0
 */

export interface DeleteTransactionsBody {
  /** @minItems 1 */
  ids: number[];
}