# validation/generated/types/updateSavingsGoalBody.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Api
 * Rayo AI Personal Finance API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateSavingsGoalBody {
  /** @minLength 1 */
  name?: string;
  /** @minimum 0.01 */
  targetAmount?: number;
  /** @minimum 0 */
  currentAmount?: number;
  deadline?: string;
}